+++ src/components/trace/index.tsx
@@ -4,11 +4,17 @@ import { ObservationTree } from "./ObservationTree";
 import { ObservationPreview } from "./ObservationPreview";
 import { TracePreview } from "./TracePreview";
 
+import Header from "@/src/components/layouts/header";
+import { Badge } from "@/src/components/ui/badge";
+import { TraceAggUsageBadge } from "@/src/components/token-usage-badge";
+import Decimal from "decimal.js";
+import { type RouterOutput } from "@/src/utils/types";
+
 export function Trace(props: {
   observations: Array<Observation & { traceId: string }>;
   trace: Trace;
   scores: Score[];
-  projectIdForManualScores?: string;
+  projectId: string;
 }) {
   const router = useRouter();
   const currentObservationId = router.query.observation as string | undefined;
@@ -52,7 +58,7 @@ export function Trace(props: {
           <ObservationPreview
             observations={props.observations}
             scores={props.scores}
-            projectIdForManualScores={props.projectIdForManualScores}
+            projectId={props.projectId}
             currentObservationId={currentObservationId}
           />
         )}
@@ -64,6 +70,52 @@ export function Trace(props: {
           scores={props.scores}
           currentObservationId={currentObservationId}
           setCurrentObservationId={setCurrentObservationId}
+        />
+      </div>
+    </div>
+  );
+}
+
+export function TracePage({
+  trace,
+}: {
+  trace: RouterOutput["traces"]["byId"];
+}) {
+  const totalCost = trace.observations.reduce(
+    (acc, o) => {
+      if (!o.price) return acc;
+
+      return acc
+        ? acc.plus(o.price ? o.price : new Decimal(0))
+        : new Decimal(0).plus(o.price ? o.price : new Decimal(0));
+    },
+    undefined as Decimal | undefined,
+  );
+
+  return (
+    <div className="flex flex-col overflow-hidden xl:container lg:h-[calc(100vh-100px)] xl:h-[calc(100vh-50px)]">
+      <Header title="Trace Detail" />
+      <div className="flex gap-2">
+        {trace.externalId ? (
+          <Badge variant="outline">External ID: {trace.externalId}</Badge>
+        ) : null}
+        {trace.userId ? (
+          <Badge variant="outline">User ID: {trace.userId}</Badge>
+        ) : null}
+        <TraceAggUsageBadge observations={trace.observations ?? []} />
+        {totalCost ? (
+          <Badge variant="outline">
+            Total cost: {totalCost.toString()} USD
+          </Badge>
+        ) : undefined}
+      </div>
+      <div className="mt-5 flex-1 overflow-hidden border-t pt-5">
+        <Trace
+          key={trace.id}
+          trace={trace}
+          scores={trace.scores}
+          projectId={trace.projectId}
+          observations={trace.observations ?? []}
         />
       </div>
     </div>
